# openapi/pulse-api.yaml
openapi: 3.0.0
paths:
...
  /auth/register:
    post:
      description: "Deja la sesión abierta: devuelve el access token y planta la cookie de refresco."
      operationId: register
      parameters:
        - name: x-pulse-client
          required: true
          in: header
          schema:
            type: string
        - name: X-Pulse-Client
          in: header
          description: "`native` para recibir el refresh token en el cuerpo."
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDto"
        "409":
          description: Ya hay una cuenta con ese correo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      summary: Crear una cuenta
      tags: &a1
        - Sesión
  /auth/login:
    post:
      operationId: login
      parameters:
        - name: x-pulse-client
          required: true
          in: header
          schema:
            type: string
        - name: X-Pulse-Client
          in: header
          description: "`native` para recibir el refresh token en el cuerpo."
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDto"
        "401":
          description: Correo o contraseña incorrectos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      summary: Iniciar sesión
      tags: *a1
  /auth/refresh:
    post:
      description: "Rota el refresh token: el anterior queda anulado en el mismo canje. Si uno ya usado vuelve a aparecer se cierran todas las sesiones de esa cuenta, porque solo se copia lo que se ha robado."
      operationId: refresh
      parameters:
        - name: x-pulse-client
          required: true
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDto"
        "401":
          description: Refresh token ausente, caducado o ya usado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      summary: Renovar el access token
      tags: *a1
  /auth/logout:
    post:
      description: "Anula el refresh token y borra la cookie. Es idempotente: sin sesión abierta también responde 204."
      operationId: logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshDto"
      responses:
        "204":
          description: ""
      summary: Cerrar la sesión actual
      tags: *a1
  /auth/forgot-password:
    post:
      description: Responde 204 exista o no la cuenta. Decir cuál de las dos convertiría este formulario en una forma de averiguar qué correos están registrados.
      operationId: forgotPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordDto"
      responses:
        "204":
          description: Si esa cuenta existe, el enlace va de camino.
      summary: Pedir un enlace para restablecer la contraseña
      tags: *a1
  /auth/reset-password:
    post:
      description: "Cierra todas las sesiones abiertas de esa cuenta: si la contraseña la tenía alguien más, su sesión también cae."
      operationId: resetPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordDto"
      responses:
        "204":
          description: Contraseña cambiada. Hay que volver a iniciar sesión.
        "401":
          description: El enlace no es válido, ya se usó o caducó.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      summary: Fijar una contraseña nueva con el enlace recibido
      tags: *a1
  /auth/verify-email/request:
    post:
      operationId: requestEmailVerification
      parameters: []
      responses:
        "204":
          description: Enviado, o innecesario porque ya estaba verificado.
      security:
        - BearerAuth: []
      summary: Pedir el correo de confirmación
      tags: *a1
  /auth/verify-email:
    post:
      operationId: verifyEmail
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailDto"
      responses:
        "204":
          description: Correo confirmado.
        "401":
          description: El enlace no es válido, ya se usó o caducó.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      summary: Confirmar el correo con el enlace recibido
      tags: *a1
  /auth/me:
    get:
      operationId: getMe
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileDto"
        "401":
          description: Falta el access token, o no es válido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security:
        - BearerAuth: []
      summary: Perfil de la sesión actual
      tags: *a1
...
info:
  title: Pulse Music API
  description: |-
    Catálogo y reproducción de la biblioteca de Pulse.

    La autenticación es global: salvo lo marcado como público —`/health`, `/catalog`,
    `/tracks/{id}/stream`, `/media/covers/{file}` y el propio `/auth`—, toda ruta exige sesión.

    `/tracks/{id}/stream` es público en el sentido de que no pide un Bearer: un `<audio src>` no
    puede mandar cabeceras. En su lugar exige la firma que emite `/catalog` en cada `streamUrl`,
    válida 24 horas. Sin ella responde 403, y la forma de conseguir una nueva es volver a pedir
    el catálogo.

    Lo que hay bajo `/admin` mantiene el catálogo y pide rol `ADMIN`. Se entra de dos maneras:
    con la sesión de una cuenta administradora, o con la cabecera `X-Pulse-Admin-Token`, que
    queda reservada a los procesos automatizados. Un credencial válido pero sin rol recibe 403,
    no 401: no hay nada que renovar.

    Lo que hay bajo `/auth` abre y renueva sesiones. El access token es un JWT de vida corta que
    viaja en `Authorization: Bearer`; el refresh token dura semanas y sale en una cookie
    `httpOnly` que solo se manda a `/v1/auth`. Un cliente sin cookies —la app móvil— lo pide en
    el cuerpo con la cabecera `X-Pulse-Client: native`.

    Toda ruta cuelga del prefijo `/v1`, ya incluido en la URL del servidor.
  version: 0.1.0
  contact: {}
...
servers:
  - url: http://localhost:4000/v1
    description: Servidor
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Pulse-Admin-Token
      description: "Credencial de servicio para procesos sin persona detrás, igual a `PULSE_ADMIN_TOKEN` en el servidor. No es para el navegador: desde la web se entra con una cuenta de rol ADMIN."
    BearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: "Access token devuelto por `/auth/login`. Caduca a los 15 minutos: renuévalo en `/auth/refresh`."
  schemas:
...
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          example: ana@ejemplo.com
          description: Se guarda en minúsculas.
        password:
          type: string
          example: una-contraseña-larga
          minLength: 10
          maxLength: 128
          description: "Nunca se almacena: solo su hash argon2id."
        displayName:
          type: string
          example: Ana
          description: Si falta, se usa la parte del correo anterior a la arroba.
      required:
        - email
        - password
    Role:
      type: string
      enum:
        - USER
        - ADMIN
    UserProfileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: ana@ejemplo.com
        displayName:
          type: string
          example: Ana
        avatarUrl:
          type: string
          nullable: true
        role:
          allOf:
            - $ref: "#/components/schemas/Role"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - displayName
        - avatarUrl
        - role
        - createdAt
    SessionDto:
      type: object
      properties:
        accessToken:
          type: string
          description: "JWT de acceso. Va en la cabecera `Authorization: Bearer`."
        expiresIn:
          type: number
          example: 900
          description: Segundos de vida que le quedan al access token.
        user:
          $ref: "#/components/schemas/UserProfileDto"
        refreshToken:
          type: string
          description: "Solo para clientes nativos, que lo piden con la cabecera `X-Pulse-Client: native`. En el navegador este campo no viene: el refresh viaja en una cookie `httpOnly`."
      required:
        - accessToken
        - expiresIn
        - user
    ApiErrorDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example:
            - title must be longer than or equal to 1 characters
        error:
          type: string
          example: Bad Request
      required:
        - statusCode
        - message
        - error
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: ana@ejemplo.com
        password:
          type: string
          example: una-contraseña-larga
      required:
        - email
        - password
    RefreshDto:
      type: object
      properties:
        refreshToken:
          type: string
          description: "Solo para clientes nativos. En el navegador se ignora: manda la cookie."
    ForgotPasswordDto:
      type: object
      properties:
        email:
          type: string
          example: ana@ejemplo.com
      required:
        - email
    ResetPasswordDto:
      type: object
      properties:
        token:
          type: string
          description: El token que viaja en el enlace del correo.
        password:
          type: string
          minLength: 10
          maxLength: 128
      required:
        - token
        - password
    VerifyEmailDto:
      type: object
      properties:
        token:
          type: string
          description: El token que viaja en el enlace del correo.
      required:
        - token
...